+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- CASM-4427: Implement a prototype to have granular query from main and sub ConfigMap
+- CASM-3981: Add S3 artifacts and loftsman manifests to the Product Catalog Schema
+
 ### Dependencies
 - Bump `google-auth` from 2.21.0 to 2.22.0 (#258)